% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The training data frame.}

\item{cl}{A vector of true class value of \code{train}.}

\item{k_nn}{An int representing the number of neighbors in knn.}

\item{k_cv}{An int representing the number of folds in cv.}
}
\value{
A list of predicted classes \code{class} and,
the mean rate of cross-validation misclassification error \code{cv-err}.
}
\description{
This function does Cross-Validation k-Nearest Neighbors prediction.
}
\examples{
train <- na.omit(my_penguins) \%>\% dplyr::select(body_mass_g, bill_length_mm,
                                                bill_depth_mm,
                                                flipper_length_mm)
cl <- na.omit(my_penguins) \%>\% dplyr::select(species)
my_knn_cv(train, cl, 5, 5)

}
\keyword{prediction}
