% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear regression function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula object input used to fit \code{data}.}

\item{data}{Numeric data frame to be fitted by \code{formula}.}
}
\value{
A table of linear model parameter of \code{formula} based on,
\code{data}, includes \code{Estimate}, \code{Std. Error},
\code{t value}, and \code{Pr(>|t|)}.
}
\description{
This function does linear regression according to formula and data.
}
\examples{
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
\keyword{inference}
