% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_pow.R
\name{my_pow}
\alias{my_pow}
\title{Power function}
\usage{
my_pow(x, power = 2)
}
\arguments{
\item{x}{Numeric input to be raised to the power of \code{power}.}

\item{power}{Numeric input for the power that \codee{x} will be raised to,
default to \code{2}.}
}
\value{
Numeric representing \code{x} raised to the power of \code{power}.
}
\description{
This functions raises input to a power.
}
