% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(train, k_cv)
}
\arguments{
\item{train}{A training data frame from my_penguins.}

\item{k_cv}{An int representing the number of folds in cv.}
}
\value{
A number of the mean rate of cross-validation misclassification.
}
\description{
This function does Cross-Validation Random Forest prediction.
}
\examples{
my_rf_cv(rf_train_df, 5)

}
\keyword{prediction}
