% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numberic input}

\item{alternative}{Character string specifying the alternative hypothesis.
Only accept \code{two.sided}, \code{less}, or \code{greater}}

\item{mu}{}
}
\value{

}
\description{
This function performs a one sample t-test
}
\examples{


}
