% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t_test}
\alias{my_t_test}
\title{T Test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.
only accepts strings \code{"two.sided}, \code{"less"}, or \code{"greater"}.}

\item{mu}{Numeric representing the null hypothesis value of the mean.}
}
\value{
List with elements
\code{test_stat}: the numeric test statistic,
\code{df}: the degrees of freedom,
\code{alternative}: the value of the parameter \code{alternative},
\code{p_val}: the numeric p-value.
}
\description{
This function performs a one sample t-test.
}
\examples{
\dontrun {
data_1 <- c(1:25)
my_t_test(data_1, "two.sided", 0)

data_2 <- c(1:100)
my_t_test(data_2, "less", 49)
}

}
\keyword{inference}
