% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Gapminder data.}
\format{
An object of class \code{data.frame} with 189 rows and 1 columns.
}
\usage{
my_gapminder
}
\description{
Excerpt of the Gapminder data on life expectancy, GDP per capita, and
population by country.
This data and documentation come from the \code{gapminder} package,
available as \code{\link[gapminder]{gapminder}}.
}
\keyword{datasets}
