% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_gapminder.R
\docType{data}
\name{my_gapminder}
\alias{my_gapminder}
\title{Gapminder Data}
\format{a data frame with 1704 observations and 6 variables:
\describe{
\item{country}{country}
\item{continent}{continent}
\item{year}{year}
\item{lifeExp}{life expectancy, in years}
\item{pop}{population}
\item{gdpPercap}{GDP Per Capita, in dollars}
}}
\source{
Bryan, Jennifer (2017) An excerpt of the data available at Gapminder.org.
}
\usage{
my_gapminder
}
\description{
The famous \code{\link[gapminder]{gapminder}} dataset taken from the
\code{gapminder} package to add datasets to our R package.
}
\keyword{datasets}
