% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame}

\item{cl}{Input data frame.}

\item{k_nn}{Integer representing the number of neighbors}

\item{k_cv}{Integer representing the number of folds}
}
\value{
a list with objects, containing \code{class} representing a vector of
the predicted class $\hat{Y}_{i}$ for all observations, and \code{cv_err} which
is a numeric with the cross-validation misclassification error
}
\description{
This functions will predict output class by using input data and gives out
cross-validation error
}
