% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbor cross-validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{A data frame of training set cases. Include only covariates of
interest to be used in predicting class.}

\item{cl}{A vector including the true class values of the training data}

\item{k_nn}{An integer representing the number of neighbors considered.}

\item{k_cv}{An integer representing the number of folds used in
cross-validation.}
}
\value{
A list with the following elements:
\code{"class"}: a vector of the predicted class for all observations,
and \code{"cv_err"}: a numeric with the cross-validation misclassification
error.
}
\description{
This function uses k-nearest neighbor analysis to generate predicted
classifications for an input dataset based on input covariates. Also uses
cross-validation to compute average misclassification rate in order to
assist in the selection of an optimal value of k.
}
\examples{
penguins <- tidyr::drop_na(my_penguins)
my_knn_cv(penguins[, 3:6], penguins$species, 1, 5)

}
\keyword{prediction}
