% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{data frame input.}

\item{cl}{true class value of the training data.}

\item{k_nn}{numeric input of the number of neighbors.}

\item{k_cv}{numeric input of the number of folds.}
}
\value{
A list with the predicted class, and a numeric output of the cross
  validation error.
}
\description{
This function uses k-nearest neighbors cross validation to predict the output
class.
}
\examples{
my_knn_cv(train = my_gapminder[, c(4, 6)], cl = my_gapminder$country, k_nn = 5,
          k_cv = 5)

}
\keyword{prediction}
