% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Output class predictor}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The data frame inputted.}

\item{cl}{The true class of the training data.}

\item{k_nn}{Numeric input representing the number of neighbors.}

\item{k_cv}{Numeric input representing the number of folds.}
}
\value{
A list with a vector of predicted classes and a numeric value with
the predicted misclassification rate.
}
\description{
This function predicts the species of penguins based on the input data.
}
\examples{
 my_knn_cv(data_use, data_clean$species, k_nn = 1, k_cv = 5)
 
}
