% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{my_knn_cv}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame to be validated}

\item{cl}{true class value of the training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
a list with objects, which has class:  ^Y_i for all observations,
& cv_err: a numeric with the cross-validation mis-classification error.
}
\description{
This function performs a k-Nearest Neighbors Cross-Validation in R.
set.seed numeric input to make the sampling reproductive within same parameter.
}
\examples{
data("my_penguins")
my_penguins <- na.omit(my_penguins)

my_knn_cv(train = my_penguins[, 3:6]
                        , cl = my_penguins[, 1],  k_nn = 5, k_cv = 5)
# my_penguins[,3:6] stands for columns #3 to #6 in the data-set:
#   "bill_length_mm",
#   "bill_depth_mm",
#   "flipper_length_mm",
#   "body_mass_g".
# column 1 is species

}
\keyword{Cross-Validation}
\keyword{Function}
\keyword{K-Nearest}
\keyword{Neighbors}
