% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k_Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame.}

\item{cl}{true class value of the training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
a list contains a vector of the predicted class for all
observations: \code{"class"}, and a numeric of the cross-validation
misclassification error: \code{"cv_err"}.
}
\description{
This function classifies obervations and computes the misclassification error.
}
\examples{
penguins_df <- na.omit(my_penguins)
penguins_cl <- penguins_df$species
penguins_df <- penguins_df \%>\%
  dplyr::select(bill_length_mm, bill_depth_mm, flipper_length_mm, body_mass_g)
my_knn_cv(penguins_df, penguins_cl, 1, 5)
}
\keyword{prediction}
