% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My K-Nearnest Neighbors Cross Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data.frame object where columns are the variables used to
predict classification}

\item{cl}{Data.frame object where ncols == 1. This column is the true
classifications of the nodes}

\item{k_nn}{The number of nearest neighbors to look at for nearest neighbors}

\item{k_cv}{The number of folds to use in cross validation}
}
\value{
A List containing:
\itemize{
\item \code{class} - vector with all classes
\item \code{cv_err} - numeric with the average CV misclassification error
}
}
\description{
Uses k-fold cross validation to perform k-nearest neighbors.
}
\details{
IMPORTANT NOTE: K-Nearest Neighbors requires no NA values in the
data. Therefore, this function removes NA values from the passed in
data.frames. If the length of the vector output$class is less than the
number of rows in the input 'train,' this may be the reason.
}
\examples{
my_knn_cv(train = my_penguins[,c("bill_length_mm",
                                 "bill_depth_mm",
                                 "flipper_length_mm",
                                 "body_mass_g")],
          cl = my_penguins[,c("species")],
          k_nn = 1,
          k_cv = 5)

}
