% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{\code{my_knn_cv} Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{an input data frame.}

\item{cl}{true class value of the training data.}

\item{k_nn}{integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
a \code{list} with objects:

\code{class}: a vector of the predicted class \code{Y_hat} for all observations,
\code{cv_err}: a numeric with the cross-validation misclassification error.
}
\description{
This function predict output class using covariates.
}
\examples{
my_knn_cv(my_penguins[, 3:6], my_penguins$species, 5, 5)

}
\keyword{prediction}
