% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Cross Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{Vector of true class values of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A vector of the predicted class for all observations.

A numeric with the cross-validation misclassification error.
}
\description{
Function that performs k-Nearest Neighbors Cross Validation.
}
\examples{
train <- iris[1:4]
cl <- iris$Species
my_knn_cv(train, cl, 5, 10)

}
