% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of the training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list containing a vector of the predicted class for all observations, and
a numeric with the cross-validation misclassification error.
}
\description{
This function predicts output class using covariates within the input data
}
\examples{


}
\keyword{prediction}
