% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{knn_cv}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{a numeric input data frame}

\item{cl}{the true class value of the training data}

\item{k_nn}{a numeric representing the desired number of nearest neighbors}

\item{k_cv}{a numeric representing the number of folds to be used for k-fold
cross validation}
}
\value{
A list of two objects where the first object are the predicted outcomes
based on the input data and the second object is the average
cross-validation misclassification error rate
}
\description{
This function takes input data, the desired number of nearest neighbors,
the number of folds to be used for k-fold cross validation and performs
k-fold cross validation from the given input data
}
\examples{
data("my_penguins")
data <- my_penguins[c("bill_length_mm",
                  "bill_depth_mm",
                  "flipper_length_mm",
                  "body_mass_g",
                  "species")]
data <- as.data.frame(data)
# remove rows containing NA values
data_noNA <- na.omit(data)
# split data into predictors and outcome
train <- data_noNA[, 1:4]
cl <- data_noNA[, 5]
# test function with 1-nearest neighbor and 5-fold cv
results_1nn_5cv <- my_knn_cv(train, cl, 1, 5)
# test function with 5-nearest neighbor and 5-fold cv
results_5nn_5cv <- my_knn_cv(train, cl, 5, 5)

}
\keyword{cross}
\keyword{fold,}
\keyword{k}
\keyword{nearest}
\keyword{neighbors,}
\keyword{prediction}
\keyword{validation,}
