% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{my_knn_cv function}
\usage{
my_knn_cv(k_cv, train, cl, k_nn)
}
\arguments{
\item{k_cv}{Integer representing the number of folds.}

\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}
}
\value{
a list with objects: class of speciese and CV error
}
\description{
This function predict output class \code{species} using covariates
\code{bill_length_mm}, \code{bill_depth_mm}, \code{flipper_length_mm}, and \code{body_mass_g}.
}
\examples{
train <- na.omit(my_penguins)
my_knn_cv(5, train[,3:6],train[[1]], 1)

}
\keyword{\code{prediction}}
