% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest Neighbors Cross-validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame of covariate values used to predict a class}

\item{cl}{String indicating the output class to predict}

\item{k_nn}{Numeric for the number of neighbors to use in the k-nearest neighbors algorithm}

\item{k_cv}{Numeric for the number of folds for cross-validation}
}
\value{
list with the cross-validation error and a vector of the predicted class
}
\description{
This function is used to predict a class using the k-nearest neighbors algorithm
}
\examples{
my_knn_cv(my_penguins[complete.cases(my_penguins), c(1, 3:6)], "species", 5, 5)
my_knn_cv(my_penguins[complete.cases(my_penguins), c(1, 3:5)], "species", 10, 10)

}
\keyword{prediction}
