% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The input data frame.}

\item{cl}{The true class value of training data.}

\item{k_nn}{The integer representing the number of neighbors.}

\item{k_cv}{The integer representing the number of folds.}
}
\value{
List with objects:
\describe{
\item{class}{a vector of the predicted class Y hat for all observations}
\item{df}{a numeric with the cross-validation misclassification error}
}
}
\description{
This function predicts output from existing variables
  by using k-nearest neighbors cross-validation
}
\examples{
my_knn_cv(my_iris[, -5], my_iris$Species, 5, 5)

}
\keyword{prediction}
