% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame, training data.}

\item{cl}{Vector of true class value of training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list containing a vector of the predicted class \code{class} and
a numeric with the cross-validation misclassification error, \code{cv_err}.
}
\description{
This function predicts an output class using the input data's covariates.
}
\examples{
#remove na values
train <- na.omit(my_penguins)
#remove unnecessary rows
train$island <- NULL
train$sex <- NULL
train$year <- NULL
cl <- train$species
my_knn_cv(train, cl, 1, 5)

}
\keyword{prediction}
