% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
List with objects: a vector of the predicted class Y for all
observations \code{class} and a numeric with the cross-validation
misclassification error \code{cv_err}.
}
\description{
This function predicts output class.
}
\examples{
data <- na.omit(my_penguins)
train <- data[, 3:6]
my_knn_cv(train, data$species, 1, 5)

}
\keyword{prediction}
