% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest neighbors cross validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Inpuut data frame of covariates.}

\item{cl}{True class value of your training data/covariates.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list with the following elements:
class: a vector of the predicted class for all observations
cv_error: a numeric with the cross-validation misclassification error.
}
\description{
This function predicts output class using covariates and finds CV misclassification error.
}
\examples{
train <- rbind(iris3[1:25,,1], iris3[1:25,,2], iris3[1:25,,3])
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
my_knn_cv(train, cl, 4, 5)

}
\keyword{prediction}
