% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest Neighbors Cross Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
list with the following elements:
class: a vector of the predicted class y for all observations
cv_err: a numeric with the cross-validation misclassification error
}
\description{
This function performs knn cross validation on a set of data.
}
\examples{
data <- my_penguins \%>\%
        tidyr::drop_na()
train_data <- data \%>\%
              dplyr::select(bill_length_mm, bill_depth_mm,
                     flipper_length_mm, body_mass_g)
target_class <- data$species

my_knn_cv(train_data, target_class, 1, 5)

}
\keyword{knn_cv}
