% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{is an input data frame containing the columns of data that are
used for the training part of the algorithm.}

\item{cl}{is a data frame input containing the column of predictions that
the function should reference for Mean Squared Error calculations.}

\item{k_nn}{is a numeric input indicating amount of neighbors the algorithm
should be using.}

\item{k_cv}{is a numeric input indicating amount of folds the function
should be using.}
}
\value{
A list where the first element is a vector of predictions created
for each \code{k_cv} folds, and the second element is the cross-validation
estimate calculated using Means Squared Errors taken from each fold.
}
\description{
This function runs the K-Nearest Neighbors Cross-Validation algorithm on the
the given data using the given \code{k_nn} and \code{k_cv} value.
}
\examples{
x <- c(3, 4, 5, 6, 7, 2, 3)
y <- c(10, 14, 50, 16, 27, 32, 13)
z <- c(3, 4, 4, 3, 4, 2, 1)
train <- data.frame(x, z)
cl <- data.frame(y)
my_knn_cv(train, cl, 1, 5)
my_knn_cv(train, cl, 2, 2)

}
\keyword{prediction}
