% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data frame of training set from input.}

\item{cl}{True class value of training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{integer representing the number of folds.}
}
\value{
A list with \code{class} a vector of the predicted class
for all observations, and \code{cv_err} the cross-validation misclassification error.
}
\description{
This function use k-nearest neighbbour cross-validatory classfication for trainig set.
}
\examples{
library(palmerpenguins)
penguins <- na.omit(penguins)
my_knn_cv(penguins[, 3:6], penguins$species, 1, 5)

}
\keyword{prediction}
