% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{KNN K-fold CrossValidation Algorithm}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Training data set to be given to \code{my_knn_cv}.}

\item{cl}{Class of object to be predicted by \code{my_knn_cv}.}

\item{k_nn}{Number of k nearest numbers to be used by \code{knn}.}

\item{k_cv}{Number of folds for k fold cross validation.}
}
\value{
List of predicted class vector \code{cl} and average cv error.
}
\description{
This function returns knn average cv error and predictions.
}
\examples{
my_df <- my_penguins
my_df <- my_df \%>\%
 dplyr::select(c(species, bill_length_mm, bill_depth_mm,
          flipper_length_mm, body_mass_g))
my_df$species <- as.character(my_df$species)
my_df <- na.omit(my_df)
my_df_class <- my_df \%>\% dplyr::pull(species)
my_knn_cv(train = my_df, cl = "species", k_nn = 1, k_cv = 5)
my_knn_cv(train = my_df, cl = "species", k_nn = 5, k_cv = 5)

}
\keyword{Prediction}
