% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The training data frame.}

\item{cl}{A vector of true class value of \data{train}.}

\item{k_nn}{An int representing the number of neighbors in knn.}

\item{k_cv}{An int representing the number of folds in cv.}
}
\value{
A list of predicted classes \code{class} and,
the mean rate of cross-validation misclassification error \code{cv-err}.
}
\description{
This function does Cross-Validation k-Nearest Neighbors prediction.
}
\examples{
my_knn_cv(train_df, result_df, 5, 5)

}
\keyword{prediction}
