% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The training input data frame.}

\item{cl}{True class value of \code{train}.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list of predicted \code{class} and a numeric with the
cross-validation misclassification error \code{cv-err}.
}
\description{
This function perform Cross-Validation k-Nearest Neighbors Cross-Validation.
}
\examples{
data_cleaned <- na.omit(my_penguins)

data_train <- data_cleaned[, c("bill_length_mm", "bill_depth_mm",
  "flipper_length_mm", "body_mass_g")]
  my_knn_cv(data_train, data_cleaned$species, 1, 5)
  my_knn_cv(data_train, data_cleaned$species, 5, 5)

}
\keyword{prediction}
