% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation Function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{The training input data frame.}

\item{cl}{True class value of \code{train}.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list of predicted \code{class} and a numeric with the
cross-validation misclassification error \code{cv-err}.
}
\description{
This function perform Cross-Validation k-Nearest Neighbors Cross-Validation.
}
\examples{
train <- na.omit(my_penguins) \%>\%
  dplyr::select(body_mass_g, bill_length_mm,
                bill_depth_mm,flipper_length_mm)
cl <- na.omit(my_penguins) \%>\% dplyr::select(species)
my_knn_cv(train, cl, 5, 10)

}
\keyword{prediction}
