% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-nearest-neighbor function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
a list with objects "class": a vector of the predicted class
for all observations; "cv_err": a numeric with the cross-validation
misclassification error
}
\description{
This function performs k nearest neighbors with k-fold cross validation.
}
\examples{

dat <- na.omit(my_penguins[, c("species", "bill_length_mm", "bill_depth_mm",
                               "flipper_length_mm", "body_mass_g")])
cl <- dat$species
dat <- dat[, c("bill_length_mm", "bill_depth_mm",
               "flipper_length_mm", "body_mass_g")]
my_knn_cv(dat, cl, 1, 5)

}
\keyword{prediction}
