% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of your training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list consists of a vector \code{class} comprised of prediction made
by \code{knn()} function from class package, and misclassification
rate \code{cv_err}, a value between 0 and 1 representing the
proportion of observations that were classified incorrectly.
}
\description{
This function splits the data into k groups for evaluating test set and
training sets to predict the best value referring the neighbors.
}
\examples{
penguins <- STAT302package::my_penguins
penguins2 <- penguins \%>\% drop_na()
my_cl <- penguins2 \%>\% pull(species)
result_nn1 <- my_knn_cv(penguins2[, 3:6], my_cl, 1, 5)

}
\keyword{prediction}
