% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn = 1, k_cv = 5, feedback = FALSE)
}
\arguments{
\item{train}{A data frame containing the covariates (x).}

\item{cl}{A vector containing the true classification of the training
data (y).}

\item{k_nn}{A numeric specifying the number of neighbors.}

\item{k_cv}{A numeric specifying the number of folds.}

\item{feedback}{A boolean for debugging feedback.}
}
\value{
A list comprised of two components:  \code{class} (a vector of
predicted classes for all observations) and \code{cv_err} (a numeric
cross-validation misclassification error calculated by taking the mean
}
\description{
This function applies cross-validation to a nearest neighbor algorithm.
}
\examples{
train <- as.data.frame(rbind(iris3[1:25,,1], iris3[1:25,,2], iris3[1:25,,3]))
cl <- factor(c(rep("s",25), rep("c",25), rep("v",25)))
my_knn_cv(train, cl, k_nn = 1, k_cv = 5, feedback = FALSE)
my_knn_cv(train, cl, k_nn = 1, k_cv = 5)
my_knn_cv(train, cl, k_nn = 1)
my_knn_cv(train, cl)

}
\keyword{prediction}
