% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My version of k-Nearest Neighbors Cross-Validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{a integer representing the number of neighbors}

\item{k_cv}{a integer representing the number of folds}
}
\value{
a \code{list} with following objects:

\code{class}: a vector of the predicted class for all observations

\code{cv_error}: a numeric with the cross-validation misclassification error
}
\description{
This function predict output class of input data by
K-nearest neighbor algorithm
}
\examples{
## Demonstrate using my_penguins dataset
# get training data
penguins_train <- dplyr::select(na.omit(my_penguins),
  bill_length_mm, bill_depth_mm, flipper_length_mm, body_mass_g)
# get true class
penguins_cl <- as.character(na.omit(my_penguins)$species)
my_knn_cv(penguins_train, penguins_cl, 1, 5)

}
\keyword{prediction}
