% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{Fitting Linear Models}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{cl}{true class value of your training data}

\item{k_nn}{integer representing the number of neighbors}

\item{k_cv}{integer representing the number of folds}
}
\value{
a list with objects

\item{class}{a vector of the predicted class Ŷ i for all observations}
\item{cv_err}{a numeric with the cross-validation misclassification error}
}
\description{
This function is used to predict output class species using covariates bill_length_mm,
bill_depth_mm, flipper_length_mm, and body_mass_g.
}
\examples{
library(palmerpenguins)
data(package = "palmerpenguins")
penguins_df <- na.omit(penguins)
train <- lapply(
  penguins_df[c("bill_length_mm", "bill_depth_mm", "flipper_length_mm", "body_mass_g")],
   as.numeric)
cl <- as.numeric(penguins_df$species)
nearest_1 <- my_knn_cv(train, cl, 1, 5)
nearest_5 <- my_knn_cv(train, cl, 5, 5)

}
\keyword{perdiction}
