% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{k-Nearest Neighbors Cross-Validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{An input data frame as training variables.}

\item{cl}{True class value of your training data \code{train}}

\item{k_nn}{An integer representing the number of neighbors}

\item{k_cv}{An integer representing the number of folds}
}
\value{
A list with two objects:
\enumerate{
\item \code{class}: A vector of the predicted class Y^i for all observations
\item \code{cv_err}: A numeric value with the cross-validation misclassification error
}
}
\description{
This function performs a k-nearest neighbors cross-validation.
}
\examples{
my_penguins <- na.omit(penguins[, 1:6])
train <- my_penguins[, 3:6]
my_knn_cv(train, my_penguins$species, 1, 5)

}
\keyword{prediction}
