% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross-validation function}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Input data frame.}

\item{cl}{True class value of the training data.}

\item{k_nn}{Integer representing the number of neighbors.}

\item{k_cv}{Integer representing the number of folds.}
}
\value{
A list of the predicted class and the CV misclassification rate.
}
\description{
This function performs a k-neareast neighbors cross-validation.
predicts "species" using four covariates bill_length_mm, bill_depth_mm,
flipper_length_mm, and body_mass_g.
}
\examples{
penguins_01 <- na.omit(my_penguins)
species_outcome <- dplyr::pull(penguins_01, var = "species")
penguins_01 <- dplyr::select(penguins_01, 3:6)
my_knn_cv(penguins_01, species_outcome, 5, 10)

}
\keyword{prediction}
