% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear regression}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{formula} class object, similar to \code{lm()}.}

\item{data}{Input data frame.}
}
\value{
a \code{table} similar to the coefficent table from \code{summary()} with rows for each coefficient
(including the \code{(Intercept)}!) and columns for the \code{Estimate}, \verb{Std. Error}, \verb{t value},
and \verb{Pr(>|t|)}.
}
\description{
This functions gives a t test based on the input
}
