% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting linear models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{An object of class \code{"formula"} that describes the model
to be fitted.}

\item{data}{An input data frame containing the variables to be used in the
model.}
}
\value{
A table with rows for each estimated coefficient and columns
including the estimate, standard error, t-statistic, and p-value.
}
\description{
This function is used to fit linear models in order to carry out regressions.
}
\examples{
my_lm(mpg ~ hp + wt, mtcars)

}
\keyword{inference}
