% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model fucntion}
\usage{
my_lm(form, data)
}
\arguments{
\item{form}{the input formula.}

\item{data}{The inputted dataset.}
}
\value{
A tibble containing the estimate, standard error, t value, and p
value for the intercept.
}
\description{
This function performs linear regression on given data and models.
}
\examples{
my_lm(form, my_gapminder)

}
