% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(
  formula,
  data,
  subset,
  weights,
  na.action,
  method = "qr",
  model = TRUE,
  x = FALSE,
  y = FALSE,
  qr = TRUE,
  singular.ok = TRUE,
  contrasts = NULL,
  offset,
  ...
)
}
\arguments{
\item{formula}{Formula class object that \code{lm} will use.}

\item{data}{Input data frame for \code{lm}.}
}
\value{
Table with rows for each coefficient, including "Intercept",
and columns for "Estimate", "Std. Error", "t value", and "Pr(>|t|):.
}
\description{
This function fits a linear model of data in R.
}
\examples{
my_lm(formula = lifeExp ~ gdpPercap + continent, data = my_gapminder)

}
\keyword{inference}
