% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Fitting Linear Models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{"formula"} class object (a symbolic description of
the model to be fitted).}

\item{data}{input data frame.}
}
\value{
a \code{"table"} with rows for each coefficient and columns for the
\code{"Estimate"}, \code{"Std.Error"}, \code{"t_value"},
and \code{"Pr"}.
}
\description{
This function fits a linear model for input data.
}
\examples{
# Create a random input for test
value1 <- c(182, 193, 183, 194, 189, 183)
value2 <- c(1, 3, 0, 2, 2, 3)
weight <- c(9.3, 11.1, 8.7, 10.4, 10.6, 12.3)
input <- data.frame("value1" = value1, "value2" = value2, "weight" = weight)
my_lm(weight ~ value1 + value2, input)
}
\keyword{inference}
\keyword{prediction}
