% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My OLS Regression Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula type object to use for regression.}

\item{data}{Data.frame containing the variables from which regression will
be applied.}
}
\value{
Matrix containing the estimates, standard error, t values, and
individual statistical significance for each variable used in regression.
}
\description{
My OLS Regression Function
}
\examples{
my_lm(flipper_length_mm ~ body_mass_g, data = my_penguins)

}
