% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{\code{my_lm} Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{formula} class object.}

\item{data}{an input data frame.}
}
\value{
a \code{table} with rows for each coefficient and columns for the \code{Estimate},
\code{Std. Error}, \code{t value}, and \code{Pr(>|t|)}.
}
\description{
This function fits a linear model in R.
}
\examples{
my_lm(mpg ~ hp, data = mtcars)

}
\keyword{inference}
