% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula object (model to be fitted).}

\item{data}{Input data frame.}
}
\value{
Table object containing the the estimate, standard error, t-value, and p-value
for each coefficient of the model.
}
\description{
Function that fits a linear model.
}
\examples{
my_lm(Petal.Width ~ Sepal.Length + Sepal.Width, iris)
my_lm(Petal.Length ~ Sepal.Length + Sepal.Width, iris)

}
