% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{linear model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{formula to fit the input data to}

\item{data}{input data frame}
}
\value{
A table of the parameters of the model and their respective
estimated values, standard errors, t values, and p values
}
\description{
This function takes a formula and input data and
fits a linear regression model of the data to the formula,
providing summary statistics.
}
\examples{
data(mtcars)
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
\keyword{linear-model}
\keyword{prediction}
\keyword{regression}
