% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{my_lm function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object.}

\item{data}{Input data frame.}
}
\value{
a table with four columns, estimates, std_error, t_stat, and p_value.
}
\description{
This function gives a summary of the lm object in R.
}
\examples{
my_lm(lm(formula = mpg ~ hp * wt, data = mtcars), mtcars)

}
\keyword{\code{inference}}
