% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object that indicates which classes from the data to use as for the y-variable and which to use for for the x-variable}

\item{data}{Data frame to build a linear model for}
}
\value{
Table that reports the estimate, standard error, t-value, and probability > t for each coefficient, including the intercept
}
\description{
This function is used to fit data to a linear model
}
\examples{
my_lm(lifeExp ~ gdpPercap, my_gapminder)
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
\keyword{inference}
