% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{An object of class \code{formula},
a symbolic description of the model to be fitted.}

\item{data}{A input data frame containing the variables in the model.}
}
\value{
a table with rows for each coefficient (including the \code{(Intercept)})
  and columns for the \code{Estimate}, \code{Std. Error}, \code{t value},
  and \code{Pr(>|t|)}.
}
\description{
This function fits a linear model in statistic
}
\examples{
my_lm(mpg ~ hp + wt, mtcars)

}
\keyword{inference}
