% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula class object, describes model to be fitted to.}

\item{data}{Input data frame.}
}
\value{
Table with rows for each coefficient, and columns \code{Estimate},
\code{Std. Error}, \code{t value}, and \code{Pr(>|t|)}
}
\description{
This function fits data to a linear model.
}
\examples{
my_lm(lifeExp ~ gdpPercap, my_gapminder)
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
\keyword{inference}
