% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{linear model fit function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{formula for the linear model (ie: lifeExp ~ gdpPercap + pop)}

\item{data}{data frame the formula is based on}
}
\value{
table with rows for each coefficient and columns for the Estimate,
Std. Error, t value, and Pr(>|t|)
}
\description{
This function conducts a linear model function on the given formula and data
}
\examples{
my_lm(lifeExp ~ gdpPercap + continent, data = my_gapminder)

}
\keyword{inference}
\keyword{linear}
\keyword{model,}
\keyword{prediction,}
