% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Model Fitting}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{is an input object that represents the describes
the model which needs to be fitted.}

\item{data}{is an input numeric vector which contains the
variables in the model.}
}
\value{
A table containing all coefficients created, which
include the error, estimate, t-statistic, and probability
of that coefficient occurring given the t-statistic.
}
\description{
This function fits linear models.
}
\keyword{inference,}
\keyword{prediction}
