% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear  Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Class object of \code{formula}, a symbolic
description of the model to be fitted.}

\item{data}{Data frame of input.}
}
\value{
A table with rows for each coefficient and columns
for the \code{Estimate}, \code{Std.Error}, \code{t value},
and \code{Pr(>|t|)}.
}
\description{
This function is used to fit linear models.
}
\examples{
data(mtcars)
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
