% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Function}
\usage{
my_lm(data, formula)
}
\arguments{
\item{data}{input data frame to be fitted for a model by \code{my_lm}.}

\item{formula}{input formula object to be used for regression model.}
}
\value{
data frame of beta coefficient estimates, standard error, t value,
and p-value.
}
\description{
This function uses linear regression to return the beta coefficient and
estimates, standard error, t value, and p value.
}
\examples{
my_lm(my_penguins, bill_length_mm ~ bill_depth_mm + flipper_length_mm)
my_lm(my_penguins, flipper_length_mm ~ bill_length_mm + bill_depth_mm)

}
\keyword{Inference}
\keyword{Prediction,}
