% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object to fit \code{data}.}

\item{data}{Numeric data frame to fit \code{formula}.}
}
\value{
A table of linear model parameter of \code{formula} including:
\code{data} with \code{Estimate}, \code{Std. Error},
\code{t value}, and \code{Pr(>|t|)}.
}
\description{
This function fits linear regression model.
}
\examples{
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)

}
\keyword{inference}
