% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{linear model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object, similar to lm().}

\item{data}{Input data frame.}
}
\value{
A table with rows of each coefficients and columns for the
\code{Estimate}, \code{Std. Error}, \code{t value}, and
\code{Pr(|t|)}.
}
\description{
This function fits a linear model in R
}
\examples{
data(mtcars)
test_my_lm <- my_lm(mpg ~ hp + wt, data = mtcars)
test_my_lm

}
\keyword{inference}
