% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model}
\usage{
my_lm(formula, data, pp = 0)
}
\arguments{
\item{formula}{A formula class object (same as \code{lm()}).}

\item{data}{A data frame.}

\item{pp}{A numeric specifying the scientific notation precision for the
values in the returned table.}
}
\value{
A table containing the following cols for each coefficient:
\code{Estimate} (expected change in the response due to a unit change in the
feature), \verb{Std. Error} (standard error of the estimate), \verb{t value}
(t-statistic), and \verb{Pr(>|t|)} (p-value).
}
\description{
This function fits a linear model.
}
\examples{
my_lm(mpg ~ hp + wt, data = mtcars, pp = 5)
my_lm(mpg ~ hp + wt, data = mtcars)

}
\keyword{inference}
\keyword{prediction}
