% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My version of fitting linear models}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{a \code{formula} class object.}

\item{data}{input data frame.}
}
\value{
a \code{table} similar to the coefficient table from \code{summary()},
with rows for each coefficient and columns for the \code{Estimate},
\code{Std. Error}, \code{t value}, \code{Pr(>|t|)}.
}
\description{
This function fits a linear model in R.
}
\examples{
## Demonstrate a regression using my_gapminder dataset.
my_lm(lifeExp ~ gdpPercap + gdpPercap, data = my_gapminder)

}
\keyword{inference/prediction}
