% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear model function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula class object}

\item{data}{Input data frame.}
}
\value{
A table with rows for each coefficient (including the Intercept!) and
columns for the Estimate, Std. Error, t value, and Pr(>|t|).
}
\description{
function my_lm() fits a linear model in R.
}
\examples{
my_lm(mpg ~ hp + wt,mtcars)

}
\keyword{prediction}
