% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_penguins.R
\docType{data}
\name{my_penguins}
\alias{my_penguins}
\title{Penguins Data}
\format{
The main \code{penguins} data frame is a tibble with 344 rows and 8 variables:
\describe{
\item{species}{a factor denoting penguin species (Adelie, Chinstrap and Gentoo)}
\item{island}{a factor denoting island in Palmer Archipelago, Antarctica (Biscoe, Dream or Torgersen)}
\item{bill_length_mm}{a number denoting bill length (millimeters)}
\item{bill_depth_mm}{a number denoting bill depth (millimeters)}
\item{flipper_length_mm}{an integer denoting flipper length (millimeters)}
\item{body_mass_g}{an integer denoting body mass (grams)}
\item{sex}{a factor denoting penguin sex (female, male)}
\item{year}{an integer denoting the study year (2007, 2008, or 2009)}
}
}
\source{
\url{https://cran.r-project.org/web/packages/palmerpenguins/index.html}
}
\usage{
my_penguins
}
\description{
Size measurements for adult foraging penguins near Palmer Station, Antarctica.
}
\details{
This data and documentation comes from the \code{palmerpenguins} package,
available as \code{\link[penguins]{penguins}}.
}
\keyword{datasets}
