% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{K Nearest Neighbors Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds}
}
\value{
A numeric with the cross-validation error
}
\description{
This function will predict output \code{body_mass_g} using covariates
\code{bill_length_mm}, \code{bill_depth_mm}, and \code{flipper_length_mm}. All the data
is in the dataframe penguins
}
