% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{K Nearest Neighbors Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds}
}
\value{
A numeric with the cross-validation error
}
\description{
This function will predict output body_mass_g using covariates
bill_length_mm, bill_depth_mm, and flipper_length_mm. All the data
is in the dataframe penguins
}
\examples{
\dontrun{
my_rf_cv(k = 5)
}


}
\keyword{\code{prediction}}
