% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv2.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Penguin body mass predictor}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input representing the number of folds.}
}
\value{
Numeric output representing the cv error.
}
\description{
This function predicts the penguins body mass based on physical trait inputs.
}
\examples{
 my_rf_cv(5)
 
}
