% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{my_rf_cv}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds}
}
\value{
a numeric with the cross-validation error;
}
\description{
Description
This function uses \code{bill_length_mm},
\code{bill_depth_mm}, and \code{flipper_length_mm} to predict
\code{body_mass_g} from the data \code{my_penguins} with k-fold Cross-validation
& random forest algorithms.
}
\examples{
my_rf_cv(k = 5)
my_rf_cv(k = 10)

}
\keyword{prediction}
