% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Function: my_rf_cv}
\usage{
my_rf_cv(k)
}
\description{
Function: my_rf_cv  Predict output sepal.length using covariates sepal.with
                    and petal.(width, length)
Input: a numeric number stands for number of folds
Output: a number -- the cross-valiadation error
}
\examples
{
my_rf_cv(5)
}
