% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds.}
}
\value{
a numeric with the cross-validation error
}
\description{
This function predicts \code{"body_mass_g"} in penguins data using
covariates \code{"bill_length_mm"}, \code{"bill_depth_mm"},
and \code{"flipper_length_mm"} and computes the cross-validation error.
}
\examples{
my_rf_cv(5)
}
\keyword{prediction}
