% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds.}
}
\value{
A numeric with the cross validation error.
}
\description{
This function predicts output using covariates within the data.
}
\examples{

}
\keyword{prediction}
