% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{A numeric indicating the number of folds}
}
\value{
Returns average mean squared error of the random forest model across
all k folds from k-fold cross-validation
}
\description{
This function performs k-fold cross validation on a random forest model
generated from internal data on penguins. Model trained is
body_mass_g ~ bill_length_mm + bill_depth_mm + flipper_length_mm
and number of trees is 100. User can specify the number of folds to be used.
}
\examples{
my_rf_cv(1)
my_rf_cv(5)

}
\keyword{prediction}
\keyword{random-forest}
