% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{my_rf_cv function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds.}
}
\value{
a numeric with the cross-validation error
}
\description{
This function predict predict the body mass based on characteristics of penguins
}
\examples{
my_rf_cv(5)

}
\keyword{\code{prediction}}
