% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric input indicating the number of folds for cross-validation}
}
\value{
Numeric with the cross-validation
}
\description{
This function calculates the cross-validation error of a prediction from the random forest algorithm
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
