% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross validation functions}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Numeric nunmber of folds.}
}
\value{
A numeric with the cross validation error.
}
\description{
This function predicts output with cv error based on the number of folds,
this function works with \code{my_penguins}.
}
\keyword{prediction}
