% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Integer representing the number of folds.}
}
\value{
Numeric with the cross-validation error \code{cv_err}.
}
\description{
This function predicts output \code{body_mass_g} using covariates
\code{bill_length_mm}, \code{ill_depth_mm}, and \code{flipper_length_mm}.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
