% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation Function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds}
}
\value{
number with the cross-validation error value
}
\description{
This function performs a cross validation using the random forest package
to train a model. predicting \code{body_mass_g} using the covariates of
\code{bill_length_mm}, \code{bill_depth_mm}, and \code{flipper_length_mm}.
}
\examples{
my_rf_cv(4)

}
\keyword{Cross}
\keyword{RandomForest,}
\keyword{Validation}
