% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k, cl, train)
}
\arguments{
\item{k}{is a numeric input used to specify the amount of folds the function
will use for the algorithm.}

\item{cl}{is a data frame input containing the column of predictions that
the function should reference for Mean Squared Error calculations.}

\item{train}{is an input data frame containing the columns of data that are
used for the training part of the algorithm.}
}
\value{
A numeric value indicating the cross-validation estimate calculated
while running the random forest cross-validation on the given data.
}
\description{
This function runs the Random Forest K-fold Cross-Validation algorithm on the
the given data using the given \code{k} value.
}
\keyword{prediction}
