% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest algorithm}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{The number of folds for k fold cross validation for \code{my_rf_cv}.}
}
\value{
Numeric of average mean squared error from \code{my_rf_cv}
predictions.
}
\description{
This function returns the average mean squared error for k folds of
crossvalidation for the randomforest algorithm.
}
\examples{
my_rf_cv(2)
my_rf_cv(10)

}
\keyword{Prediction}
