% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{random forest function}
\usage{
my_rf_cv(train, k_cv)
}
\arguments{
\item{train}{input data frame}

\item{k_cv}{integer representing the number of folds}
}
\value{
a numeric with the cross-validation error
}
\description{
This function performs random forest algorithm with k-fold cross validation.
}
\examples{
dat <- na.omit(my_penguins[, c("body_mass_g", "bill_length_mm",
                               "bill_depth_mm", "flipper_length_mm")])
my_rf_cv(dat, 2)

}
\keyword{prediction}
