% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Number of folds}
}
\value{
Average mean squared error across all \code{k} folds.
}
\description{
This function splits the data into k groups for evaluating test set and
training sets to predict the best value by random forest algorithm
}
\examples{
my_rf_cv(5)


}
\keyword{prediction}
