% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(k = 5, feedback = FALSE)
}
\arguments{
\item{k}{A numeric specifying the number of folds.}

\item{feedback}{A boolean for debugging feedback.}
}
\value{
A numeric specifying the average mean squared error (MSE) across
all folds.
}
\description{
This function applies cross-validation to a random forest algorithm on the
penguins data set.
}
\examples{
my_rf_cv(5, TRUE)
my_rf_cv(5)
my_rf_cv()

}
\keyword{prediction}
