% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random forest cross-validation function}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{number of folds.}
}
\value{
A numeric with the cross-validation error.
}
\description{
This function performs a random forest cross-validation.
predicts "body_mass_g" using covariates bill_length_mm, bill_depth_mm,
flipper_length_mm.
}
\examples{
my_rf_cv(5)

}
\keyword{prediction}
