% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test method}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".
Otherwise, this function will throw an informative error.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements including the numeric test statistic \code{t_test},
the degrees of freedom \code{df}, the value of the parameter \code{alternative},
and the the numeric p-value \code{p_val}.
}
\description{
This functions gives a t test based on the input
}
\examples{
my_t.test(x = rnorm(100), alternative = "less", mu = 0 )
my_t.test(x = rnorm(100), alternative = "two.sided", mu = 0 )
my_t.test(x = rnorm(100), alternative = "greater", mu = 0 )

}
\keyword{t.test}
