% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
Only accepts input values of \code{"two.sided"}, \code{"less"},
or\code{"greater"}.}

\item{mu}{Numeric input indicating the null hypothesis value of the mean.}
}
\value{
A list with the following elements:
\code{"test_stat"}: the numeric test statistic,
\code{"df"}: the degrees of freedom,
\code{"alternative"}: the value of the parameter \code{"alternative"},
and \code{"p_val"}: the numeric p-value.
}
\description{
This function performs a one-sample t-test on a numeric vector of input data.
}
\examples{
my_t.test(1:10000, alternative = "two.sided", mu = 4900)

}
\keyword{inference}
