% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test Function}
\usage{
my_t.test(x, alternative = c("two.sided", "less", "greater"), mu)
}
\arguments{
\item{x}{Numeric vector the data is saved into.}

\item{alternative}{Character string equal to "two.sided", "less", or
"greater".}

\item{mu}{Numeric input that xhat will be compared to.}
}
\value{
Returns a list of the t-value, degrees of freedom, alternative value,
and p-value.
}
\description{
This function runs a single or double tailed t test on provided data.
}
\examples{
my_t.test(x, "two.sided", 0)
  
}
