% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{The numeric vector of data given to conduct the t-test. No default
value.}

\item{alternative}{The alternative hypothesis, a string of either "two.sided",
"less", or "greater". Used to determine which t-test to conduct. No default
value.}

\item{mu}{The predicted mean, a numeric value to calculate the test statistic.
No default value.}
}
\value{
A list of the t-value, the degrees of freedom, the alternative used,
and the P-value.
}
\description{
This function calculates P-value using the t-distribution hypothesis test.
}
\examples{
my_t.test(c(1, 2, 2), alternative = "less", 3)

}
\keyword{hypothesis}
\keyword{inference}
\keyword{test}
