% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector input of data to be subjected to a t test by \code{my_t.test}.}

\item{alternative}{Character string input specifying the alternative hypothesis for
\code{my_t.test} is being run. This should only accept \code{"two.sided"}, \code{"less"}, or
\code{"greater"}. Otherwise, your function should throw an informative error.}

\item{mu}{Numeric input indicating the null hypothesis of the mean that \code{my_t.test}
will compare the mean to.}
}
\value{
List with elements:
"test_stat": the numeric test statistic.
"df": the degrees of freedom.
"alternative": the value of the parameter \code{alternative}.
"p_val": the numeric p-value.
}
\description{
This function runs a t test on a vector of numbers.
}
\examples{
x <- as.vector(subset(data.frame(my_gapminder), select = c("lifeExp")))
alternative <- c("two.sided", "less", "greater")
mu <- 60

my_t.test(x, alternative = "two.sided", mu)

my_t.test(x, alternative = "less", mu)

my_t.test(x, alternative = "greater", mu)

}
\keyword{inference}
