% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test Function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
Only accept \code{"two.sided"}, \code{"less"} or \code{"greater"}.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list with numeric test statistic \code{"test_stat"},
the degrees of freedom \code{"df"},
the value of the parameter alternative \code{"alternative"},
and the numeric p-value \code{"p_val"}.
}
\description{
This function performs one-sample t test for both two-sided or one-sided.
}
\examples{
\dontrun{
x <- rnorm(10, mean = 0.1, sd = 1)
my_t.test(x, "two.sided", 0)
my_t.test(x, "greater", 0)
}
}
\keyword{t-test}
