% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My Student's T-Test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector containing data to perform the test on.}

\item{alternative}{String with the type of t-test to use (two.sided,
less, greater).}

\item{mu}{Numeric with the expected mean of the data \code{x}.}
}
\value{
List with the test stat (\code{test_stat}), df (\code{df}),
alternative (\code{alternative}), and p values (\code{p_val}).
}
\description{
This function performs a student's t-test on the inputted data.
}
\examples{
my_t.test(my_penguins$bill_depth_mm, alternative = "two.sided", mu = 15)
my_t.test(my_penguins$body_mass_g, alternative = "less", mu = 4000)

}
