% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis. This should only accept `
"two.sided"`, `"less"`, or `"greater".}

\item{mu}{Number indicating the null hypothesis value of the mean.}
}
\value{
A list containing the numeric test statistic, degrees of freedom, value of the \code{alternative}
parameter, and the numeric p-value.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
my_t.test(1:10, "two-sided", 5)

}
\keyword{inference}
