% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of the input data}

\item{alternative}{A character representing the hypothesis
test we want to compute: "two.sided", "less", or "greater}

\item{mu}{A number indicating the null hypothesis value of the mean}
}
\value{
A list containing the test statistic, the degrees of freedom, the
hypothesis test, and the computed p value
}
\description{
This function takes input data, a hypothesis test, a null hypothesis
value, and returns a list of the computed t statistic, the degrees
of freedom, the hypothesis test used, and the computed p value
}
\examples{
x <- rnorm(100, mean = 0, sd = 1)
my_t.test(x, "less", 0)
my_t.test(x, "greater", 0)
my_t.test(x, "two.sided", 0)

}
\keyword{inference}
\keyword{statistical}
\keyword{t-test,}
\keyword{test,}
