% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.}

\item{mu}{Number indicating the null hypothesis value of the mean}
}
\value{
A list with four elements, test_stat, df, alternative, and p_value.
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(rbinom(100, size = 7, prob = 0.1), "greater", 0.9)
my_t.test(rbinom(100, size = 7, prob = 0.1), "less", 0.9)
my_t.test(rbinom(100, size = 7, prob = 0.1), "two.sided", 0.9)

}
\keyword{\code{inference}}
