% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector to perform t-test on}

\item{alternative}{String that specifies whether alternative hypothesis is "two.sided", "less", or "greater"}

\item{mu}{Numeric that indicates the null hypothesis value of the mean}
}
\value{
test_stat Numeric test statistic

df Numeric degrees of freedom

alternative String value of the parameter "alternative"

p_val Numeric p-value
}
\description{
This function is used for performing a one-sample t-test
}
\examples{
my_t.test(c(1:100), "two.sided", 50)
my_t.test(c(1:100), "greater", 50)
my_t.test(c(1:100), "less", 50)

}
\keyword{inference}
