% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numberic vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.
Only accept \code{"two.sided"}, \code{"less"}, or \code{"greater"}.}

\item{mu}{Numberic input indicating the null hypothesis value of the mean.}
}
\value{
List with elements:
\describe{
\item{test_stat}{the numeric test statistic}
\item{df}{the degrees of freedom}
\item{alternative}{the value of the parameter alternative}
\item{p_val}{the numeric p-value}
}
}
\description{
This function performs a one sample t-test
}
\examples{
my_t.test(x, "two.sided", 12)

}
