% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis,
only accepts "two.sided", "less", or "greater".}

\item{mu}{Number indicating the null hypothesis value of the mean.}
}
\value{
List containing the numeric test statistic \code{test_stat}, numeric
degrees of freedom \code{df}, string value of parameter \code{alternative},
and numeric p-value \code{p_val}.
}
\description{
This function takes an input dataset, and performs a null hypothesis t-test.
}
\examples{
my_t.test(1:10, "two.sided", mu = 1)
my_t.test(1:10, "greater", mu = 0.1)
my_t.test(1:10, "less", mu = 10)

}
\keyword{inference}
