% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{Character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".
Otherwise, your function should throw an informative error.}

\item{mu}{Number indicating the null hypothesis value of the mean.}
}
\value{
List with elements: the numeric test statistic \code{test_stat},
the degrees of freedom \code{df},
the value of the parameter \code{alternative},
the numeric p-value \code{p_val}.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
set.seed(302)
x <- rnorm(10, mean = 1, sd = 1)
# two-sided t-test
my_t.test(x, alternative = "two.sided", mu = 0.5)
# one-sided t-test -- upper tail
my_t.test(x, alternative = "greater", mu = 0.5)
# one-sided t-test -- lower tail
my_t.test(x, alternative = "less", mu = 0.5)

}
\keyword{inference}
