% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data.}

\item{alternative}{A character string that specifies alternative hypothesis
must be "less", "greater", or "two.sided".}

\item{mu}{Numeric input indicating the null hypothesis value of the mean.}
}
\value{
A list with the following elements:
test_stat: the numeric test statistic
df: the degrees of freedom
alternative: the value of the parameter \code{alternative}
p_val: the numeric p-value.
}
\description{
This function performs one sample t-test.
}
\examples{
# Creating a random vector
set.seed(302)
random_vec <- rnorm(10, mean = 1.5, sd = 1)

# Performing one sample two sided t-test
my_t.test(x = random_vec, "two.sided", mu = 3)

# Performing one sample one sided t-test
my_t.test(x = random_vec, "greater", mu = 3)

}
\keyword{inference}
