% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test}
\usage{
my_t.test(x, alternative = "two.sided", mu)
}
\arguments{
\item{x}{is a numeric input vector representing the data for
the t-test.}

\item{alternative}{is a string input used to distinguish what
t-test to run, a two-sided, less, or greater, defaults to
\code{"two.sided"}}

\item{mu}{is a numeric input indicating the given hypothesized
mu value.}
}
\value{
A list of t-test values that have been calculated, this
includes the t-statistic, degrees of freedom, the
\code{alternative}, and the p-value.
}
\description{
This function runs specified t-test type on given data using
the given hypothesized mu.
}
\keyword{inference}
