% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-Test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Non-empty numeric vector of data values.}

\item{alternative}{String to specify the alternative hypothesis,
must be one of \code{two.sided}, \code{greater} or \code{less}.}

\item{mu}{Number indicating the true value of the mean.}
}
\value{
List to show the value of the \code{t-statistic}, \code{the degrees} of
freedom, the \code{alternative} option and the \code{p-value} of the test.
}
\description{
This function performs one sample t-test on vectors of data.
}
\examples{
a <- rbinom(100, size = 10, prob = 0.2)
my_t.test(a, "greater", 0.8)

}
\keyword{inference}
