% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data to be tested.}

\item{alternative}{A string specifying the alternative hypothesis,
only accept "two.sided", "less", or "greater".}

\item{mu}{The null hypothesis of the mean of \data{x}.}
}
\value{
A list of t-test parameters, includes \code{test_stat},
\code{df}, \code{alternative}, and \code{p_val}.
}
\description{
This function does one sample t-test.
}
\examples{
my_t.test(my_gapminder$lifeExp, "two.sided", 60)
my_t.test(my_gapminder$lifeExp, "less", 60)
my_t.test(my_gapminder$lifeExp, "greater", 60)

}
\keyword{inference}
