% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T Test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric vector of data}

\item{alternative}{A string specifying types of t test wanted, with choices
"two.sided", "less" or "greater".}

\item{mu}{a number indicating the null hypothesis.}
}
\value{
a list including: the numeric test statistic, the degrees of freedom,
the type of t test given, and the numeric p-value.
}
\description{
This function performs t test on given data
}
\examples{
my_t.test(my_gapminder$lifeExp, "two.sided", 60)
my_t.test(my_gapminder$lifeExp, "less", 60)
my_t.test(my_gapminder$lifeExp, "greater", 60)

}
\keyword{inference}
