% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{T-test}
\usage{
my_t.test(x, alternative = "two.sided", mu = 0)
}
\arguments{
\item{x}{A numeric vector of data}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater".
Otherwise, your function should throw an informative error.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list consisting following elements:
\code{test_stat} : The numeric test statistics
\code{df} : The degree of freedom
\code{alternative} : The value of parameter \code{alternative}
\code{p_val} : The numeric p value
}
\description{
This function performs t-test for given value.
}
\examples{
x <- rnorm(10, mean = 0, sd = 1)
my_t.test(x, alternative = "two.sided", mu = 0)

}
\keyword{inference}
