% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test}
\usage{
my_t.test(x, alternative = "two.sided", mu = 0)
}
\arguments{
\item{x}{Numeric vector of data comprising sample values.}

\item{alternative}{String specifying the alternative hypothesis, can be
one of three values: less, two.sided, or greater.}

\item{mu}{The numeric mean of the null hypothesis.}
}
\value{
A list comprised of the following: \code{test_stat} (the numeric
test statistic), \code{df} (the degrees of freedom), \code{alternative}
(a string representing the alternative hypothesis), and \code{p_val} (the
numeric p-value).
}
\description{
This function performs a one sample t-test.
}
\examples{
my_t.test(rnorm(100, mean = 0, sd = 1), alternative = "less", mu = 0)
my_t.test(rnorm(100, mean = 0, sd = 1), alternative = "greater")
my_t.test(rnorm(100, mean = 0, sd = 1))

}
\keyword{inference,}
\keyword{statistical}
\keyword{t-test}
\keyword{test,}
