% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My version of t-test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values to be tested.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of "two.sided", "greater" or "less".}

\item{mu}{a numeric input indicating the null hypothesis value of the mean.}
}
\value{
a \code{list} with following element:

\code{test_stat}: the numeric test statistic.

\code{df}: the degrees of freedom.

\code{alternative}: the value of the parameter \code{alternative}.

\code{p_val}: the numeric p-value.
}
\description{
This function performs a one sample t-test in R.
}
\examples{
my_t.test(my_gapminder$lifeExp, "two.sided", 60)
my_t.test(my_gapminder$lifeExp, "greater", 60)
my_t.test(my_gapminder$lifeExp, "less", 60)

}
\keyword{inference}
