% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{Student's t-Test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept \code{"two.sided"}, \code{"less"}, or \code{"greater"}.}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
a list with elements:

\item{test_stat}{the numeric test statistic}
\item{df}{the degrees of freedom}
\item{alternative}{the value of the parameter alternative}
\item{p_val}{the numeric p-value}
}
\description{
This function performs a one sample t-test in R.
}
\examples{
trial <- rnorm(30, 4, 7)
t.test(trial, alternative = "less", mu = 4)

}
\keyword{inference}
