% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t-test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or "greater". Otherwise,
the function throw an informative error.}

\item{mu}{A number indicating the null hypothesis value of the mean.}
}
\value{
A list with elements,which includes
\code{test_stat}: the numeric test statistic.
\code{df}: the degrees of freedom.
\code{alternative}: the value of the parameter \code{alternative}.
\code{p_val}: the numeric p-value.
}
\description{
my_t.test() function performs a one sample t-test in R.
}
\examples{
data <- rnorm(100, mean=25, sd=10)
my_t.test(data, "two.sided", 15)
my_t.test(data, "less", 15)
my_t.test(data, "greater", 15)


}
\keyword{inference}
