% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{t test function}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{a numeric vector of data.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept "two.sided", "less", or greater".}

\item{mu}{a number indicating the null hypothesis value of the mean.}
}
\value{
A list with the numeric test statistic, the degrees of freedom,
the value of the parameter \code{alternative}, the numeric p-value.
}
\description{
This function can perform a one sample t test in R.
}
\examples{
lifeExp <- my_gapminder$lifeExp
my_t.test(lifeExp, "two.sided", 60)

}
\keyword{inference}
