% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{Sample t-test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.}

\item{alternative}{A string that specify the alternative hypothesis.}

\item{mu}{A number that indicates the null hypothesis value of the mean.}
}
\value{
A list with the following elements: test statistics, degree of freedom
       parameter of alternative, numeric p-value.
}
\description{
This function does the same as the built-in function t.test().
}
\examples{
my_t_test(c(2, 5, 3), "less", 2)

}
\keyword{t-test}
