% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T-Test function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{Vector of numerics given to \code{my_t_test}.}

\item{alternative}{Character string of test condition i.e. "two sided,"
"less," or "greater."}

\item{mu}{numeric input of mu value for t test.}
}
\value{
List of test statistic, df, input character string, and p value.
}
\description{
This function calculates t test statistic, degrees of freedom, and
p value given vector of numerics and character string of test condition.
}
\examples{
sample_data <- rnorm(1000, 12, 5)
my_t_test(sample_data, "two.sided", 10)
my_t_test(sample_data, "less", 10)

}
\keyword{inference}
