% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t_test.R
\name{my_t_test}
\alias{my_t_test}
\title{T-test Function}
\usage{
my_t_test(x, alternative, mu)
}
\arguments{
\item{x}{A numeric vector of data.
value.}

\item{alternative}{a character string specifying the alternative hypothesis.
This should only accept \code{"two.sided"}, \code{"less"}, or \code{"greater"}. Otherwise,
the function should throw an informative error.}

\item{mu}{A numeric indicating the null hypothesis value of the mean.}
}
\value{
A list of numeric including the t-value, the degrees of freedom, the
alternative, and the p-value.
}
\description{
This function run a one sample t-test.
}
\examples{
my_t_test(my_gapminder$lifeExp, "two.sided", 60)
my_t_test(my_gapminder$lifeExp, "less", 60)
my_t_test(my_gapminder$lifeExp, "greater", 60)

}
\keyword{Test}
\keyword{hypothesis}
\keyword{inference}
